p14) -10ms: large u16 array much faster than large u32 array for memoization:
test b14 ... bench:  25,124,867 ns/iter (+/- 2,385,614) u16 array
test b14 ... bench:  35,440,213 ns/iter (+/- 1,750,331) u32 array
test b14 ... bench:  35,818,913 ns/iter (+/- 2,088,310) u32 array
test b14 ... bench:  25,371,235 ns/iter (+/- 1,390,509) u16 array

+++ src/collatz.rs
@@ -23,7 +23,7 @@ impl Iterator for Collatz {
     }
 }
 
-pub fn count_collatz(n: u32, lengths: &mut [u32; 500_000]) -> u32 {
+pub fn count_collatz(n: u32, lengths: &mut [u16; 500_000]) -> u16 {
     if n < 500_000 && lengths[n as usize] > 0 {
         lengths[n as usize]
     } else {
@@ -44,7 +44,7 @@ pub fn count_collatz(n: u32, lengths: &mut [u32; 500_000]) -> u32 {
 
 pub fn longest_collatz_memo(highest: u32) -> u32 {
     let mut max_length = 0;
-    let mut lengths: [u32; 500_000] = [0; 500_000];
+    let mut lengths: [u16; 500_000] = [0; 500_000];
     let mut cause = 0;
     for i in 1..highest + 1 {
         let length = count_collatz(i, &mut lengths);
